1. 使用jwt构造身份cookie 2. 添加后台接口的权限验证 3. format code

+++ src/main/java/easeplan/netease/sales/controller/CartController.java
@@ -1,12 +1,15 @@
 package easeplan.netease.sales.controller;
 
 import easeplan.netease.sales.domain.CartItem;
+import easeplan.netease.sales.service.IAuthService;
 import easeplan.netease.sales.service.ICartService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
@@ -19,47 +22,75 @@ import java.util.List;
 @Controller
 public class CartController {
     @Autowired
+    IAuthService authService;
+    @Autowired
     ICartService cartService;
 
     @RequestMapping(path = "/cart", method = RequestMethod.GET)
     public ModelAndView cartPage(
-            @RequestHeader(value = "Referer", required = false, defaultValue = "/") String referer
-    ) {
-        ModelAndView mav = new ModelAndView("cart");
-        List<CartItem> cartItems = cartService.getCartItemList();
-        mav.addObject("cartItems", cartItems);
-        int deletedItemCount = cartService.deleteDeletedItems();
-        mav.addObject("deletedItemCount", deletedItemCount);
-
-        try {
-            if (!new URI(referer).getPath().startsWith("/cart")) {
-                mav.addObject("referer", referer);
+            @RequestHeader(value = "Referer", required = false, defaultValue = "/") String referer,
+            @CookieValue(value = "identity", required = false) String identity,
+            HttpServletResponse response
+    ) throws IOException {
+        if (authService.isBuyer(identity)) {
+            ModelAndView mav = new ModelAndView("cart");
+            List<CartItem> cartItems = cartService.getCartItemList();
+            mav.addObject("cartItems", cartItems);
+            int deletedItemCount = cartService.deleteDeletedItems();
+            mav.addObject("deletedItemCount", deletedItemCount);
+            try {
+                if (!new URI(referer).getPath().startsWith("/cart")) {
+                    mav.addObject("referer", referer);
+                }
+            } catch (URISyntaxException e) {
             }
-        } catch (URISyntaxException e) {
+            return mav;
+        } else {
+            response.sendRedirect("/");
+            return null;
         }
-        return mav;
     }
 
     @RequestMapping(value = "/api/cart/{id}", method = RequestMethod.POST)
     @ResponseBody
     public void addItem(
-            @PathVariable int id
-    ) {
-        cartService.addItem(id);
+            @PathVariable int id,
+            @CookieValue(value = "identity", required = false) String identity,
+            HttpServletResponse response
+    ) throws IOException {
+        if (authService.isBuyer(identity)) {
+            cartService.addItem(id);
+        } else {
+            response.sendError(401);
+        }
     }
 
     @RequestMapping(value = "/api/cart/{id}", method = RequestMethod.PUT)
     @ResponseBody
     public void updateItemAmount(
             @PathVariable int id,
-            @RequestParam int num
-    ) {
-        cartService.updateItemAmount(id, num);
+            @RequestParam int num,
+            @CookieValue(value = "identity", required = false) String identity,
+            HttpServletResponse response
+    ) throws IOException {
+        if (authService.isBuyer(identity)) {
+            cartService.updateItemAmount(id, num);
+        } else {
+            response.sendError(401);
+        }
     }
 
     @RequestMapping(value = "/api/cart/{id}", method = RequestMethod.DELETE)
     @ResponseBody
-    public void deleteItem(@PathVariable int id) {
-        cartService.deleteItem(id);
+    public void deleteItem(
+            @PathVariable int id,
+            @CookieValue(value = "identity", required = false) String identity,
+            HttpServletResponse response
+    ) throws IOException {
+        if (authService.isBuyer(identity)) {
+            cartService.deleteItem(id);
+        } else {
+            response.sendError(401);
+        }
     }
 }
